added piano experiment, some reorganisation
I changed '_cgi-bin' to 'include' and grep+sed'd everything in the
repository accordingly. I was not aware of `.git/index`, which was corrupted by
this. This was fixed by deleting the file and using `git reset`.

I also removed the `index.html` file which was causing issues with
GitHub pages.

Finally, and most importantly, I started and pretty much completed an
entire experiment! Czech it out: https://blieque.github.io/piano/